back-api/requirements.txt
fastapi>=0.121.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
sqlalchemy[mypy]>=2.0.44
psycopg2-binary>=2.9.9
alembic>=1.17.1
pyjwt[crypto]>=2.10.1
bcrypt>=5.0.0
python-multipart>=0.0.6
fastapi-error-map>=0.9.8
email-validator>=2.0.0
pgvector>=0.3.0
openai>=1.0.0